src/requirements.txt
fastapi==0.104.1
aiofiles==23.2.1
aiohttp==3.9.1
aiomysql==0.2.0
aioredis==2.0.1
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-multipart==0.0.6
PyJWT==2.8.0
loguru==0.7.2
tortoise-orm==0.20.0
aiosmtplib==3.0.1
cryptography==42.0.5